+++ src/tests/test_parse_nlog_log4net_syntax/Program.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace test_parse_nlog_log4net_syntax
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+        }
+    }
+}